[FE] design: 헤더 포지션 수정 및 모달 컴포넌트 위치 조정
