+++ frontend/src/component/button/FriendPageButtons.js
@@ -1,38 +1,23 @@
 import OpenModalButton from '@component/button/OpenModalButton';
-import ModalComponent from '@component/modal/ModalComponent';
-import FriendWaitList from '@component/contents/FriendWaitList';
-import FriendSearch from '@component/contents/FriendSearch';
+import BasicButton from '@component/button/BasicButton';
 
 const FriendPageButtons = () => {
   const FriendWaitBtn = OpenModalButton({
     text: '> WAITING',
-    classList: 'btn btn-no-outline-hover fs-8',
+    classList: 'btn btn-no-outline-hover fs-7',
     modalId: '#friendWaitModal',
   });
   const FriendAddBtn = OpenModalButton({
     text: '> ADD',
-    classList: 'btn btn-no-outline-hover fs-8',
+    classList: 'btn btn-no-outline-hover fs-7',
     modalId: '#friendAddModal',
   });
-  const FriendWaitModal = ModalComponent({
-    borderColor: 'mint',
-    title: 'WAITING',
-    modalId: 'friendWaitModal',
-    content: FriendWaitList(),
-    buttonList: [],
+  const ReloadRoomBtn = BasicButton({
+    id: 'reloadRoomBtn',
+    text: '> Reload',
+    classList: 'btn btn-no-outline-hover fs-7',
   });
-  const FriendAddModal = ModalComponent({
-    borderColor: 'pink',
-    title: 'ADD FRIEND',
-    modalId: 'friendAddModal',
-    content: FriendSearch(),
-    buttonList: [],
-  });
-
-  const BtnList = [
-    `${FriendWaitBtn} ${FriendWaitModal}`,
-    `${FriendAddBtn} ${FriendAddModal}`,
-  ];
+  const BtnList = [`${ReloadRoomBtn}`, `${FriendWaitBtn}`, `${FriendAddBtn}`];
 
   return `
   	${BtnList.join('')}
